--- src/components/nav.js
@@ -1,6 +1,0 @@
-import React from "react";
-import "../styles/layout.css";
-
-export default class Nav extends React.Component {
-    
-}
